Move Delete Analysis message box to presenter
