+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/analyses/presenter/AnalysesPresenterDefaultAppearance.java
@@ -94,4 +94,14 @@ public class AnalysesPresenterDefaultAppearance implements AnalysesView.Presente
         return "300px";
     }
 
+    @Override
+    public String analysesExecDeleteWarning() {
+        return analysesMessages.analysesExecDeleteWarning();
+    }
+
+    @Override
+    public String warning() {
+        return iplantDisplayStrings.warning();
+    }
+
 }
